Separo estilos de componenetes
