+++ src/usuarios/PantallaUsuarios.jsx
@@ -1,10 +1,8 @@
 import React, { useState, useEffect } from "react";
 import { Container, Typography } from "@material-ui/core";
-import { withStyles, makeStyles } from "@material-ui/core/styles";
 
 import Table from "@material-ui/core/Table";
 import TableBody from "@material-ui/core/TableBody";
-import TableCell from "@material-ui/core/TableCell";
 import TableContainer from "@material-ui/core/TableContainer";
 import TableHead from "@material-ui/core/TableHead";
 import TableRow from "@material-ui/core/TableRow";
@@ -16,32 +14,10 @@ import ProgresoCircular from "../components/ProgresoCircular";
 
 import ModalDeshabilitar from "../components/ModalDeshabilitar";
 import ModalEditarUsuario from "./ModalEditarUsuario";
+import { useStyles } from "../components/styles";
+import { StyledTableCell, StyledTableRow } from "../components/StyledTable";
 
 import * as AuthServerService from "../comunications/AuthServerService";
-
-const StyledTableCell = withStyles((theme) => ({
-  head: {
-    backgroundColor: theme.palette.common.black,
-    color: theme.palette.common.white,
-  },
-  body: {
-    fontSize: 14,
-  },
-}))(TableCell);
-
-const StyledTableRow = withStyles((theme) => ({
-  root: {
-    "&:nth-of-type(odd)": {
-      backgroundColor: theme.palette.action.hover,
-    },
-  },
-}))(TableRow);
-
-const useStyles = makeStyles({
-  table: {
-    minWidth: 700,
-  },
-});
 
 const PantallaUsuarios = () => {
   const classes = useStyles();
