+++ src/videos/PantallaVideos.jsx
@@ -1,6 +1,5 @@
 import React, { useState, useEffect } from "react";
 import { Container, Typography } from "@material-ui/core";
-import { withStyles, makeStyles } from "@material-ui/core/styles";
 
 import TableContainer from "@material-ui/core/TableContainer";
 import Paper from "@material-ui/core/Paper";
@@ -8,7 +7,6 @@ import Table from "@material-ui/core/Table";
 import TableHead from "@material-ui/core/TableHead";
 import TableBody from "@material-ui/core/TableBody";
 import TableRow from "@material-ui/core/TableRow";
-import TableCell from "@material-ui/core/TableCell";
 import Alert from "@material-ui/lab/Alert";
 import DoneIcon from "@material-ui/icons/Done";
 import ClearIcon from "@material-ui/icons/Clear";
@@ -18,30 +16,8 @@ import ProgresoCircular from "../components/ProgresoCircular";
 import ModalDeshabilitar from "../components/ModalDeshabilitar";
 import * as MediaServerService from "../comunications/MediaServerService";
 import ModalEditarVideo from "./ModalEditarVideo";
-
-const StyledTableCell = withStyles((theme) => ({
-  head: {
-    backgroundColor: theme.palette.common.black,
-    color: theme.palette.common.white,
-  },
-  body: {
-    fontSize: 14,
-  },
-}))(TableCell);
-
-const StyledTableRow = withStyles((theme) => ({
-  root: {
-    "&:nth-of-type(odd)": {
-      backgroundColor: theme.palette.action.hover,
-    },
-  },
-}))(TableRow);
-
-const useStyles = makeStyles({
-  table: {
-    minWidth: 700,
-  },
-});
+import { useStyles } from "../components/styles";
+import { StyledTableCell, StyledTableRow } from "../components/StyledTable";
 
 const PantallaVideos = () => {
   const [videos, setVideos] = useState(null);
